misc/HtmlTestSuite/results/test-0653.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:dgrogan@chromium.org" rel="author" title="David Grogan"/>
<link href="https://drafts.csswg.org/css-flexbox/#intrinsic-sizes" rel="help"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

<meta content="Item with unresolvable width:100% contributes to max-content size via its content contributions." name="assert"/>

<style>
  .grandchild {
    float: left;
    width: 50px;
  }

</style>

</head><body><div data-expected-width="100" style="display: flex; flex-flow: column wrap; width: max-content; height: 100px; background: green;">
  <div data-expected-width="100" style="width: 100%; height: 50px; min-height: 0px;">
    <div class="grandchild"/>
    <div class="grandchild"/>
  </div>
</div>


<script>
  checkLayout('body &gt; div');
</script>
</body></html>
